Agent: Rename credential_repository package
credential_repository -> propagation_credentials_repository

Issue #2773
PR #2780
