+++ monkey/infection_monkey/agent_event_handlers/add_stolen_credentials_to_repository.py
@@ -1,7 +1,7 @@
 import logging
 
 from common.agent_events import CredentialsStolenEvent
-from infection_monkey.credential_repository import IPropagationCredentialsRepository
+from infection_monkey.propagation_credentials_repository import IPropagationCredentialsRepository
 
 logger = logging.getLogger(__name__)
