+++ monkey/infection_monkey/propagation_credentials_repository/i_propagation_credentials_repository.py
